fix: resolve pre-existing errors uncovered by serde fix
- Remove .unwrap() from ArkDerivationIndex::new() (returns Self, not Result)
- Cast min_signers/total_signers to u16 for FrostThreshold::new()
- Remove unused FrostSigningSession import in roast.rs

These errors were masked by earlier compilation failures (HashMap
scope + serde big-array trait bounds). Fixing those unblocked
compilation, revealing additional issues.

+++ src/protocol/frost.rs
@@ -696,7 +696,7 @@ impl FrostSigningContext {
 
         let participants_set = FrostParticipantSet::new(participants)
             .map_err(|e| ConclaveError::CryptoError(format!("FROST keygen set: {e:?}")))?;
-        let threshold = FrostThreshold::new(min_signers, total_signers)
+        let threshold = FrostThreshold::new(min_signers as u16, total_signers as u16)
             .map_err(|e| ConclaveError::CryptoError(format!("FROST keygen thresh: {e:?}")))?;
 
         Ok(FrostKeyPackage {
